Abstract class MediumPermanentFormatter
